Refactor backend structure and enhance configuration management
- Removed .env.sample and added .env for environment variable management.
- Updated .gitignore to exclude .env files and added Terraform-related entries.
- Introduced Dockerfile for containerizing the FastAPI application.
- Created Makefile for managing deployment tasks and environment checks.
- Refactored application structure to include a new file processing module.
- Updated dependencies in poetry.lock and pyproject.toml for compatibility.
- Enhanced logging in S3 client for better error tracking.
- Removed outdated Terraform configurations and added new ones for ECR and API Gateway.
- Change Lambda to use Container based executions
