--- backend/.aws/terraform/main.tf
@@ -1,208 +1,0 @@
-provider "aws" {
-  region = var.aws_region
-}
-
-resource "aws_iam_role" "lambda_exec" {
-  name = "lambda_exec_role"
-  assume_role_policy = jsonencode({
-    Version = "2012-10-17",
-    Statement = [{
-      Action = "sts:AssumeRole",
-      Effect = "Allow",
-      Principal = {
-        Service = "lambda.amazonaws.com"
-      }
-    }]
-  })
-}
-
-# Add CloudWatch Logs policy to Lambda execution role
-resource "aws_iam_role_policy_attachment" "lambda_logs" {
-  role       = aws_iam_role.lambda_exec.name
-  policy_arn = "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole"
-}
-
-resource "aws_lambda_function" "fastapi" {
-  filename         = "${path.module}/../../dist/package.zip"
-  function_name    = var.lambda_function_name
-  role             = aws_iam_role.lambda_exec.arn
-  handler          = "app.main.handler"
-  source_code_hash = filebase64sha256("${path.module}/../../dist/package.zip")
-  runtime          = "python3.11"
-  timeout          = 15
-  memory_size      = 256
-  publish          = true
-
-  environment {
-    variables = {
-      LOG_LEVEL = "INFO"
-      PYTHONPATH = "/var/task"
-      S3_BUCKET_NAME = aws_s3_bucket.file_storage.id
-    }
-  }
-
-  depends_on = [
-    aws_iam_role_policy_attachment.lambda_logs
-  ]
-}
-
-# Create CloudWatch Log Group for API Gateway
-resource "aws_cloudwatch_log_group" "api_gateway" {
-  name              = "/aws/apigateway/fastapi-api"
-  retention_in_days = 30
-}
-
-# Create IAM role for API Gateway CloudWatch logging
-resource "aws_iam_role" "api_gateway_cloudwatch" {
-  name = "api-gateway-cloudwatch-role"
-  assume_role_policy = jsonencode({
-    Version = "2012-10-17"
-    Statement = [
-      {
-        Action = "sts:AssumeRole"
-        Effect = "Allow"
-        Principal = {
-          Service = "apigateway.amazonaws.com"
-        }
-      }
-    ]
-  })
-}
-
-# Add CloudWatch Logs policy to API Gateway role
-resource "aws_iam_role_policy" "api_gateway_cloudwatch" {
-  name = "api-gateway-cloudwatch-policy"
-  role = aws_iam_role.api_gateway_cloudwatch.id
-
-  policy = jsonencode({
-    Version = "2012-10-17"
-    Statement = [
-      {
-        Effect = "Allow"
-        Action = [
-          "logs:CreateLogGroup",
-          "logs:CreateLogStream",
-          "logs:PutLogEvents"
-        ]
-        Resource = "*"
-      }
-    ]
-  })
-}
-
-resource "aws_apigatewayv2_api" "api" {
-  name          = "fastapi-api"
-  protocol_type = "HTTP"
-}
-
-resource "aws_apigatewayv2_integration" "lambda" {
-  api_id             = aws_apigatewayv2_api.api.id
-  integration_type   = "AWS_PROXY"
-  integration_uri    = aws_lambda_function.fastapi.invoke_arn
-  integration_method = "POST"
-  payload_format_version = "2.0"
-  passthrough_behavior = "WHEN_NO_MATCH"
-}
-
-# Add root route for handling requests to the base URL
-resource "aws_apigatewayv2_route" "root" {
-  api_id    = aws_apigatewayv2_api.api.id
-  route_key = "ANY /"
-  target    = "integrations/${aws_apigatewayv2_integration.lambda.id}"
-}
-
-resource "aws_apigatewayv2_route" "default" {
-  api_id    = aws_apigatewayv2_api.api.id
-  route_key = "ANY /{proxy+}"
-  target    = "integrations/${aws_apigatewayv2_integration.lambda.id}"
-}
-
-resource "aws_apigatewayv2_stage" "dev" {
-  api_id      = aws_apigatewayv2_api.api.id
-  name        = "$default"
-  auto_deploy = true
-
-  access_log_settings {
-    destination_arn = aws_cloudwatch_log_group.api_gateway.arn
-    format = jsonencode({
-      requestId    = "$context.requestId"
-      httpMethod  = "$context.httpMethod"
-      status      = "$context.status"
-      responseLength = "$context.responseLength"
-    })
-  }
-}
-
-resource "aws_lambda_permission" "allow_api_gateway" {
-  statement_id  = "AllowAPIGatewayInvoke"
-  action        = "lambda:InvokeFunction"
-  function_name = aws_lambda_function.fastapi.function_name
-  principal     = "apigateway.amazonaws.com"
-  source_arn    = "${aws_apigatewayv2_api.api.execution_arn}/*/*"
-}
-
-# Create S3 bucket for file storage
-resource "aws_s3_bucket" "file_storage" {
-  bucket = var.s3_bucket_name
-}
-
-# Enable versioning for the bucket
-resource "aws_s3_bucket_versioning" "file_storage_versioning" {
-  bucket = aws_s3_bucket.file_storage.id
-  versioning_configuration {
-    status = "Enabled"
-  }
-}
-
-# Add S3 bucket policy to allow Lambda access
-resource "aws_s3_bucket_policy" "file_storage_policy" {
-  bucket = aws_s3_bucket.file_storage.id
-
-  policy = jsonencode({
-    Version = "2012-10-17"
-    Statement = [
-      {
-        Sid       = "AllowLambdaAccess"
-        Effect    = "Allow"
-        Principal = {
-          AWS = aws_iam_role.lambda_exec.arn
-        }
-        Action = [
-          "s3:PutObject",
-          "s3:GetObject",
-          "s3:DeleteObject",
-          "s3:ListBucket"
-        ]
-        Resource = [
-          aws_s3_bucket.file_storage.arn,
-          "${aws_s3_bucket.file_storage.arn}/*"
-        ]
-      }
-    ]
-  })
-}
-
-# Add S3 access policy to Lambda execution role
-resource "aws_iam_role_policy" "lambda_s3_access" {
-  name = "lambda_s3_access_policy"
-  role = aws_iam_role.lambda_exec.id
-
-  policy = jsonencode({
-    Version = "2012-10-17"
-    Statement = [
-      {
-        Effect = "Allow"
-        Action = [
-          "s3:PutObject",
-          "s3:GetObject",
-          "s3:DeleteObject",
-          "s3:ListBucket"
-        ]
-        Resource = [
-          aws_s3_bucket.file_storage.arn,
-          "${aws_s3_bucket.file_storage.arn}/*"
-        ]
-      }
-    ]
-  })
-}
