--- backend/.aws/terraform/variables.tf
@@ -1,14 +1,0 @@
-variable "aws_region" {
-  description = "AWS region"
-  type        = string
-}
-
-variable "s3_bucket_name" {
-  description = "Name of the S3 bucket"
-  type        = string
-}
-
-variable "lambda_function_name" {
-  description = "Name of the Lambda function"
-  type        = string
-}
