+++ backend/.env
@@ -1,0 +1,5 @@
+AWS_REGION=us-east-1
+S3_BUCKET_NAME=smart-file-search-bucket
+LAMBDA_NAME=smart-file-search-lambda
+ECR_REPOSITORY=smart-file-search-repository
+
