--- backend/.env.sample
@@ -1,12 +1,0 @@
-AWS_REGION=us-east-1
-TF_VAR_aws_region=${AWS_REGION}
-
-S3_BUCKET=smart-file-search-bucket
-TF_VAR_s3_bucket_name=${S3_BUCKET}
-
-LAMBDA_NAME=smart-file-search-api
-TF_VAR_lambda_function_name=${LAMBDA_NAME}
-
-AWS_ACCESS_KEY_ID=<your-access-key-id>
-AWS_SECRET_ACCESS_KEY=<your-secret-access-key>
-
