+++ backend/app/core/config.py
@@ -2,6 +2,6 @@ import os
 
 class Settings:
     AWS_REGION = os.getenv("AWS_REGION")
-    S3_BUCKET = os.getenv("S3_BUCKET")
+    S3_BUCKET_NAME = os.getenv("S3_BUCKET_NAME")
 
 settings = Settings()
