+++ backend/pyproject.toml
@@ -1,30 +1,19 @@
-[project]
+[tool.poetry]
 name = "smart-file-search-backend"
 version = "0.1.0"
-description = ""
-authors = [
-    {name = "feliperooke",email = "felipe@rooke.com.br"}
-]
-readme = "README.md"
-requires-python = ">=3.11,<3.13"
-dependencies = [
-    "fastapi (>=0.115.12,<0.116.0)",
-    "uvicorn (>=0.34.0,<0.35.0)",
-    "mangum (>=0.19.0,<0.20.0)",
-    "boto3 (>=1.34.41,<1.34.70)",
-    "aioboto3 (>=12.4.0,<13.0.0)",
-    "python-multipart (>=0.0.20,<0.0.21)",
-    "markitdown[all] @ git+https://github.com/microsoft/markitdown.git#subdirectory=packages/markitdown",
-]
+description = "FastAPI application running on AWS Lambda"
+authors = ["Felipe Rooke <felipe@rooke.com.br>"]
+package-mode = false
 
-[tool.poetry]
-packages = [
-    { include = "app", from = "." }
-]
-
-[build-system]
-requires = ["poetry-core>=2.0.0,<3.0.0"]
-build-backend = "poetry.core.masonry.api"
+[tool.poetry.dependencies]
+python = "^3.12"
+fastapi = "^0.109.0"
+mangum = "^0.17.0"
+uvicorn = "^0.27.0"
+boto3 = "^1.34.41"
+aioboto3 = "^12.4.0"
+markitdown = {version = ">=0.1.0", extras = ["docx", "pdf"]}
+python-multipart = "^0.0.20"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.5"
@@ -36,5 +25,6 @@ testpaths = ["app/"]
 python_files = ["test_*.py"]
 addopts = "-v --cov=app --cov-report=term-missing"
 
-[tool.poetry-plugin-lambda-build]
-package-artifact-path = "dist/package.zip"
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
